Improve rpc types
