+++ packages/univo/src/realtime.ts
@@ -1,242 +1,6 @@
-import WS from "ws"; // Probs make a peer dep?
-import { WebSocket } from "partysocket";
-import type { ErrorEvent } from "partysocket/ws";
-
-import { http } from "./client";
+import { IndexerRpc, NodeRpc } from "./rpc";
+import type { Transport } from "./transport";
 import { assert, createLogger, hexToNumber, iife, isHexEqual, mutex, retry } from "./utils";
-
-/**
- * Socket -----------------------------------------------------------------------------------------------------------------------------------
- */
-
-type SocketOptions = {
-	/**
-	 * An event listener to be called when the WebSocket connection's readyState changes to OPEN;
-	 * this indicates that the connection is ready to send and receive data. Note that this function
-	 * can be called multiple times over the lifetime of a socket as dropped connections are initialised
-	 */
-	onOpen?: () => Promise<void> | void;
-	/**
-	 * An event listener to be called when the WebSocket connection's readyState changes to CLOSED.
-	 */
-	onClose?: () => Promise<void> | void;
-	/**
-	 * An event listener to be called when an error occurs
-	 */
-	onError?: (event: ErrorEvent) => Promise<void> | void;
-	/**
-	 * An event listener to be called when a message is received from the server
-	 */
-	onMessage?: (event: MessageEvent) => Promise<void> | void;
-};
-
-function createSocket(url: `wss://${string}`, opts: SocketOptions) {
-	const ws = new WebSocket(url, [], { WebSocket: WS });
-
-	// No support for `binaryType` of "blob" in Bun yet so we set it to "arraybuffer" https://github.com/partykit/partykit/issues/774
-	ws.binaryType = "arraybuffer";
-
-	if (opts.onOpen) ws.onopen = opts.onOpen;
-	if (opts.onError) ws.onerror = opts.onError;
-	if (opts.onClose) ws.onclose = opts.onClose;
-	if (opts.onMessage) ws.onmessage = opts.onMessage;
-
-	return ws;
-}
-
-/**
- * Transport -----------------------------------------------------------------------------------------------------------------------------------
- */
-
-type TransportOptions = {
-	/** Logs are emitted based on the environment LOG_LEVEL. Set `quiet: true` to surpress all logs. */
-	quiet?: boolean;
-};
-
-type Transport = {
-	/**
-	 * Performs as JSON RPC request and returns the result
-	 */
-	request(opts: { method: string; params: any[] }): Promise<any>;
-	/**
-	 * Subscribes to specific event types and returns a function to unsubscribe.
-	 */
-	subscribe(param: "newHeads", handler: (message: any) => void): Promise<{ unsubscribe(): Promise<void> }>;
-};
-
-function defineTransport(url: string, opts: TransportOptions = {}): Transport {
-	if (!url.startsWith("wss://")) {
-		throw new Error("Websocket connections must start with `wss://`");
-	}
-
-	const logger = createLogger({ quiet: opts.quiet ?? false });
-
-	let id = 0;
-
-	// Every request gets an identifier based on `id` above. This map represents that request in-flight
-	// and it's corresponding callback handler to invoke when a response is received on the connection.
-	const requests = new Map<number, (data: any) => void>();
-
-	// This is a key value store for a given subscription id to a given param
-	const subscriptions = new Map<string, { param: string; latest: number }>();
-
-	// For each param we keep a set of handlers to notify when we receive a response
-	const params = new Map<string, Set<(data: any) => void>>();
-
-	const socket = createSocket(url as "wss://", {
-		async onError(cause) {
-			logger.error(new Error("Socket error", { cause }));
-		},
-
-		async onOpen() {
-			if (subscriptions.size === 0) return;
-
-			// If the socket connection is reinitalised this function will be called multiple times. When that
-			// happens we need to re-initialise all underlying subscriptions on the new connection.
-
-			const params: string[] = [];
-
-			for (const [id, subscription] of subscriptions) {
-				params.push(subscription.param);
-
-				// This isn't strictly needed for correctness. The fact that the websocket connection was dropped
-				// likely means the node automatically dropped our subscriptions. So we can fire-and-forget this.
-
-				request({ method: "eth_unsubscribe", params: [id] }).catch(() => {
-					logger.warn("Failed to cancel old susbcription");
-				});
-
-				subscriptions.delete(id);
-			}
-
-			for (const param of params) {
-				request({ method: "eth_subscribe", params: [param] }).then((id) => {
-					subscriptions.set(id, { param, latest: Date.now() });
-				});
-			}
-		},
-
-		async onMessage(event) {
-			const data = JSON.parse(event.data);
-
-			if (data.method === "eth_subscription") {
-				const subscription = subscriptions.get(data.params.subscription);
-
-				if (subscription === undefined) {
-					return await request({ method: "eth_unsubscribe", params: [data.params.subscription] }).catch(() => {
-						logger.warn("Failed to unsubscribe stale subscription");
-					});
-				}
-
-				const handlers = params.get(subscription.param);
-
-				if (handlers === undefined) {
-					return await request({ method: "eth_unsubscribe", params: [data.params.subscription] }).catch(() => {
-						logger.warn("Failed to unsubscribe stale subscription");
-					});
-				}
-
-				for (const handler of handlers) {
-					try {
-						handler(data.params.result);
-					} catch (cause) {
-						logger.error(new Error("Handler error", { cause }));
-					}
-				}
-
-				subscription.latest = Date.now();
-
-				return;
-			}
-
-			const handler = requests.get(data.id);
-
-			if (handler === undefined) {
-				return logger.warn(`Received unknown response for request id ${data.id}...`);
-			}
-
-			return handler(data.result);
-		},
-	});
-
-	async function request(opts: { method: string; params: any[] }) {
-		return await new Promise<any>((resolve) => {
-			const body = Object.assign({ id: id++ }, opts);
-
-			requests.set(body.id, (data) => {
-				// TODO: Could set a timeout to clean up the request to mitigate memory leaks
-				requests.delete(body.id);
-				resolve(data);
-			});
-
-			socket.send(JSON.stringify(body));
-		});
-	}
-
-	async function subscribe(param: string, handler: (messsage: any) => void) {
-		const handlers = params.get(param);
-
-		if (handlers === undefined) {
-			const handlers = new Set<(data: any) => void>().add(handler);
-
-			params.set(param, handlers);
-
-			const id = await request({ method: "eth_subscribe", params: [param] }).catch((cause) => {
-				throw new Error(`Failed to initialise subscription for param ${param}`, { cause });
-			});
-
-			subscriptions.set(id, { param, latest: Date.now() });
-		} else {
-			handlers.add(handler);
-		}
-
-		return {
-			async unsubscribe() {
-				const handlers = params.get(param);
-
-				if (handlers === undefined) {
-					return logger.warn("Param already unsubscribed...");
-				}
-
-				handlers.delete(handler);
-
-				if (handlers.size === 0) {
-					await request({ method: "eth_unsubscribe", params: [id] }).catch(() => {
-						logger.warn("Failed to unsubscribe unused subscription");
-					});
-				}
-			},
-		};
-	}
-
-	function healthcheck() {
-		const params: string[] = [];
-
-		for (const [id, subscription] of subscriptions) {
-			if (Date.now() - subscription.latest > 60 * 1000) {
-				params.push(subscription.param);
-
-				request({ method: "eth_unsubscribe", params: [id] }).catch(() => {
-					logger.warn("Failed to cancel old susbcription");
-				});
-
-				subscriptions.delete(id);
-
-				logger.warn(`Subscription ${id} failed health check, reinitialising..`);
-			}
-		}
-
-		for (const param of params) {
-			request({ method: "eth_subscribe", params: [param] }).then((id) => {
-				subscriptions.set(id, { param, latest: Date.now() });
-			});
-		}
-	}
-
-	setInterval(healthcheck, 60 * 1000);
-
-	return { request, subscribe };
-}
 
 /**
  * Blockchain -----------------------------------------------------------------------------------------------------------------------------------
@@ -402,27 +166,27 @@ const POLLING_INTERVAL_MS = 12 * 1000;
 type RealtimeOptions = {
 	/** Logs are emitted based on the environment LOG_LEVEL. Set `quiet: true` to surpress all logs. */
 	quiet?: boolean;
-	/** Endpoint to publish realtime blocks to */
-	endpoint: string;
-	/** Custom transport */
-	transport: Transport;
+	/** Connection to a blockchain node */
+	node: Transport<NodeRpc>;
+	/** Connection to a univo indexer */
+	indexer: Transport<IndexerRpc>;
 };
 
 function realtime(opts: RealtimeOptions) {
-	const client = http(opts.endpoint);
 	const log = createLogger({ quiet: opts.quiet ?? false });
 
 	const promise = iife(async () => {
-		let id = 0;
 		let pending: Head[] = [];
 
 		const [chain, latestBlock] = await Promise.all([
-			retry(opts.transport.request, [{ method: "eth_chainId", params: [] }], 2),
-			retry(opts.transport.request, [{ method: "eth_getBlockByNumber", params: ["latest", false] }], 2),
+			retry(opts.node.request, [{ method: "eth_chainId", params: [] }], 2),
+			retry(opts.node.request, [{ method: "eth_getBlockByNumber", params: ["latest", false] }], 2),
 		]);
 
+		const finalizedStartBlock = await opts.indexer.request({ method: "public_getUnfinalizedHeight", params: [chain] });
+
 		async function getBlockByHash(hash: `0x${string}`) {
-			const block = await opts.transport.request({
+			const block = await opts.node.request({
 				method: "eth_getBlockByHash",
 				params: [hash, false], // We don't need transaction receipts
 			});
@@ -444,20 +208,12 @@ function realtime(opts: RealtimeOptions) {
 					// The primary improvement to this retry method is that if the queue gets too large we might DDOS
 					// the endpoint by forcing it to load too many blocks at once. Moreover, this function doesn't run
 					// under any type of mutex so if the request takes longer than when we receive the next block we
-					// will spam requests. We need a strategy for queueing requests for new heads.
+					// will spam requests. We need a strategy for queueing requests for new heads. It's okay to drop
+					// unfinalised heads if the queue gets too big
 
 					const heads = pending.concat(head);
 
-					const response = await client.request({
-						id: id++,
-						jsonrpc: "2.0",
-						params: [heads],
-						method: "public_writeUnfinalizedHeads",
-					});
-
-					if (response.error) {
-						throw new Error(response.error.message);
-					}
+					await opts.indexer.request({ method: "public_writeUnfinalizedHeads", params: [heads] });
 
 					pending = pending.filter((head) => {
 						return !heads.some((_head) => {
@@ -472,16 +228,7 @@ function realtime(opts: RealtimeOptions) {
 			},
 			onBlockReorganised: async (head) => {
 				try {
-					const response = await client.request({
-						id: id++,
-						jsonrpc: "2.0",
-						params: [head],
-						method: "public_deleteReorganisedHead",
-					});
-
-					if (response.error) {
-						throw new Error(response.error.message);
-					}
+					await retry(opts.indexer.request, [{ method: "public_deleteReorganisedHead", params: [head] }], 2);
 				} catch {
 					log.warn("Failed to write reorganised head");
 				}
@@ -489,21 +236,16 @@ function realtime(opts: RealtimeOptions) {
 		});
 
 		// TODO
-		// This starts from the indexer state. As the chain finalizes we iterate over all the unfinalized
+		// This starts from the indexer state. As the chain finalizes we iterate over all the finalized
 		// blocks less than or equal to the finalized chain and deliver them to the finalized endpoint.
 		// If successful we can drop all blocks sent from the unfinalized chain
 
-		const unfinalized = defineBlockchain({
+		const indexer = defineBlockchain({
 			getBlockByHash,
 			quiet: opts.quiet ?? false,
 		});
 
-		const finalized = defineBlockchain({
-			getBlockByHash,
-			quiet: opts.quiet ?? false,
-		});
-
-		await opts.transport.subscribe("newHeads", async (head) => {
+		await opts.node.subscribe("newHeads", async (head) => {
 			try {
 				await latest.reconcile({
 					chain,
@@ -518,7 +260,7 @@ function realtime(opts: RealtimeOptions) {
 
 		async function polling() {
 			try {
-				const block = await opts.transport.request({
+				const block = await opts.node.request({
 					method: "eth_getBlockByNumber",
 					params: ["finalized", false], // We don't need transaction receipts
 				});
@@ -529,7 +271,7 @@ function realtime(opts: RealtimeOptions) {
 
 				for (const head of latest.chain) {
 					if (hexToNumber(head.number) <= hexToNumber(block.number)) {
-						await finalized.reconcile(head);
+						await indexer.reconcile(head);
 					}
 				}
 			} catch {
@@ -541,12 +283,8 @@ function realtime(opts: RealtimeOptions) {
 	});
 
 	promise.catch(() => {
-		log.error(`Failed to initialise realtime client for endpoint ${opts.endpoint}`);
+		log.error(`Failed to initialise realtime client for indexer ${opts.indexer}`);
 	});
 }
 
-/**
- * Exports -----------------------------------------------------------------------------------------------------------------------------------
- */
-
-export { realtime, defineTransport };
+export { realtime };
